Error handling and reporting in the modification framework.
Message popups are shown when an error occurs during the configuration of a filter and during the execution of a filter to inform the user of possible unexpected results. Additionally, we always print a stack trace about the error that occurred.

Errors with regard to loading available filters are printed to the error console and the filter is skipped as an option in the 'add filter' dialog.