ETM: Fitness replaying Reusing Alignments
****Fixed bugs related:
- Fixed the repair process when both the crossover and mutation take place. First it repairs the parent based on the grandparent, and then the child based on the repaired parent
- Fixed some situations where the or-joins were omitted in the repaired alignment.

****Implementation related:
- Added a alignment-validator, at least for the log part. 
- In CentralRegistry, added a function to "remove" the fitness of a tree from the cache.
- In TreeEvolutionEngine, added a function that checks if the elite population has been repaired more than a certain confidence level. If true, evaluate again.

-  

****Things to test/reconsider
(a) - TODO : It is possible to remove the loop constraint by checking a fresh move inside  the scope, therefore it would be easy to detect the  loops.
(b) -TODO: Precision
(c) -TODO/FIX: RemoveUseless can remove an entire subtree, and the parent can be lost (p.e. XOR()) so if this occurs, it returns the root. 
(d) -TODO: maybe the confidence_level should be more fuzzy, not a simple number. Does the quality is better? worse? Think of this.
(e) -TODO/FIX: What to do when removing useless behavior, all quality criteria should be the same: remove the affected move in the alignment should be enough
(f) -TODO/FIX: Related to the AStarAlgorithm: maybe it will be interesting to create a method that only changes the traces (it requires to remove the final properties of some attributes)

****Things to change/remember:
- Make the HISTORY_CACHE_SIZE dependent on the population size (size = 4*popSize)
- Remember to enable multithreading in AbstractEvolutionEnfine (right now singleThread is forced to true: test and print issues)

****Other minor changes related to the current commit:
- Some changes related to format (Ctrl+shift+F)
- Added more examples in the test plugin.